[GEOS-9938] Split the Features template module in independent parts. Removing OGCAPI dependencies from core module, move all to the dedicated one
